Fix discrepancy between declaration and definition of AAACreateAVP().
git-svn-id: https://openser.svn.sourceforge.net/svnroot/openser/trunk@3288 689a6050-402a-0410-94f2-e92a70836424

+++ modules_k/acc/diam_avp.c
@@ -107,7 +107,7 @@ AAA_AVP*  AAACreateAVP(
 	AAA_AVPFlag flags,
 	AAAVendorId vendorId,
 	char   *data,
-	unsigned int length,
+	size_t length,
 	AVPDataStatus data_status)
 {
 	AAA_AVP *avp;
